Feature ETP-1916: Improve header styles
