+++ packages/ComponentLibrary/src/components/Copilot/MessageComponents/MessageInput.tsx
@@ -1,16 +1,10 @@
-import type React from "react";
 import { useCallback, useState } from "react";
 import { InputAdornment } from "@mui/material";
 import Send from "../../../assets/icons/send.svg";
 import AttachFile from "../../../assets/icons/paperclip.svg";
 import IconButton from "../../IconButton";
 import { SearchInputWithVoice } from "../..";
-
-interface MessageInputProps {
-  onSendMessage: (message: string, files?: File[]) => void;
-  placeholder?: string;
-  disabled?: boolean;
-}
+import type { MessageInputProps } from "../types";
 
 const MessageInput: React.FC<MessageInputProps> = ({
   onSendMessage,
